transcriber_app/requirements.txt
audioread==3.0.1
av==14.4.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
coloredlogs==15.0.1
ctranslate2==4.6.0
decorator==5.2.1
faster-whisper @ https://github.com/SYSTRAN/faster-whisper/archive/refs/heads/master.tar.gz#sha256=aec3450ead9a705e96051fcaa9ae9cdb8206def46a783b038cadc88e97fa2efd
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
openai-whisper==20250625
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pycparser==2.22
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
